+++ client/src/main/java/net/oneandone/stool/client/cli/Restart.java
@@ -18,23 +18,30 @@ package net.oneandone.stool.client.cli;
 import net.oneandone.stool.client.Globals;
 import net.oneandone.stool.client.Reference;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Restart extends IteratedStageCommand {
-    private final List<String> selection;
+    private final String imageOpt;
 
-    public Restart(Globals globals, List<String> selection) {
+    public Restart(Globals globals, String imageOpt) {
         super(globals);
-        this.selection = selection;
+        this.imageOpt = imageOpt;
     }
 
     @Override
     public void doMain(Reference reference) throws Exception {
+        List<String> args;
+
         if (up(reference)) {
             new Stop(globals).doRun(reference);
         } else {
             console.info.println("Container is not running - starting a new instance.");
         }
-        new Start(globals, selection).doRun(reference);
+        args = new ArrayList<>();
+        if (imageOpt != null) {
+            args.add(imageOpt);
+        }
+        new Start(globals, args).doRun(reference);
     }
 }